Add files via upload
This Java project provides a class called FeaturesCalculator that reads binary mask data from a file, calculates the mass center of the binary mask, and saves the results to a new file. The project could be useful for image processing or computer vision applications that require the calculation of mass centers for binary masks. The class contains methods for reading the data, calculating the mass center, and saving the results, and can be easily integrated into larger Java programs.