Add files via upload
World Tour Guide is a Java project that provides a list of popular European countries and their capitals. It also provides a list of cities that are popular tourist destinations in Europe. The project uses Map, List, and Set data structures to store and retrieve the information. The project includes a method to print the list of countries and their capitals. It also includes methods to check if a specific country, city or unique city exists in the list. The project is useful for anyone who is planning to visit Europe and wants to know about popular destinations.