Add files via upload
Pascal's Triangle Generator is a Java project that generates Pascal's Triangle based on the user's input. Pascal's Triangle is a mathematical concept that shows the coefficients of binomial expansion. The project asks the user to input the height of the triangle and generates the triangle using a two-dimensional array. The project checks if the user's input is negative and if so, prints an error message. The project is useful for anyone who wants to understand and visualize Pascal's Triangle or wants to use it for mathematical calculations.